Refactor: Extract member store logic into dedicated stores
- Created faceMemberSelect.store.ts for FaceMemberSelectDialog.vue.
- Created nlEditor.store.ts for NLEditorInput.vue.
- Moved relevant member fetching and searching logic from member.store.ts to the new stores.
- Updated FaceMemberSelectDialog.vue and NLEditorInput.vue to use their respective new stores.
- Cleaned up member.store.ts by removing the extracted logic.
