+++ apps/admin/src/stores/member.store.ts
@@ -25,11 +25,7 @@ export const useMemberStore = defineStore('member', {
       sortBy: [] as { key: string; order: string }[], // Sorting key and order
     },
 
-    // State for single item operations
-    detail: {
-      item: null as Member | null,
-      loading: false, // Loading state for a single member
-    },
+
 
     // State for add operations
     add: {
@@ -173,26 +169,7 @@ export const useMemberStore = defineStore('member', {
       this._loadItems();
     },
 
-    async setCurrentItem(item: Member | null) {
-      this.detail.item = item;
-    },
 
-    async getById(id: string): Promise<Member | undefined> {
-      this.detail.loading = true;
-      this.error = null;
-      const result = await this.services.member.getById(id);
-      this.detail.loading = false;
-      if (result.ok) {
-        if (result.value) {
-          this.detail.item = result.value;
-          return result.value;
-        }
-      } else {
-        this.error = i18n.global.t('member.errors.loadById');
-        console.error(result.error);
-      }
-      return undefined;
-    },
 
     async getByIds(ids: string[]): Promise<Member[]> {
       this.list.loading = true;
@@ -215,21 +192,6 @@ export const useMemberStore = defineStore('member', {
       this.list.totalPages = 1;
     },
 
-    async searchMembers(searchQuery: string) {
-      this.list.loading = true;
-      this.error = null;
-      const result = await this.services.member.loadItems({
-        searchQuery,
-      }, 1, 10); // Search top 10
 
-      if (result.ok) {
-        this.list.items = result.value.items;
-      } else {
-        this.error = i18n.global.t('member.errors.load');
-        this.list.items = [];
-        console.error(result.error);
-      }
-      this.list.loading = false;
-    },
   },
 });
